Add files via upload
The Roll a Dice app is a simple, fun, and interactive application developed using Flutter. It allows users to roll a virtual dice and get a random number between 1 and 6, simulating the experience of rolling a physical dice. This app is perfect for board games, decision-making, or just having some fun.